requirements.txt
python-telegram-bot==20.7
aiosqlite>=0.19.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
pytest>=7.4.4